第 3 题 第3轮

+++ backend/src/main/java/com/fullStack/expenseTracker/models/Budget.java
@@ -18,11 +18,11 @@ public class Budget {
     long budgetId;
     long userId;
     Integer categoryId;
-    double amount;
+    long amount;
     int month;
     long year;
 
-    public Budget(long userId, Integer categoryId, double amount, int month, long year) {
+    public Budget(long userId, Integer categoryId, long amount, int month, long year) {
         this.userId = userId;
         this.categoryId = categoryId;
         this.amount = amount;
